Move graphics to separate module 'plotter'. Graphs for f2. Dynamic y offset based on minimum function value.

+++ functions.nim
@@ -1,7 +1,7 @@
 import math
 
-proc f1*(x: float64): float64 {.procvar.} = 
+proc f1*(x: float): float {.procvar.} = 
     result = (x + 3) * cos(x)
 
-proc f2*(x: float64): float64 {.procvar.} = 
+proc f2*(x: float): float {.procvar.} = 
     result = abs(x) * (x + 3) * cos(x)
